Refactor User and Voucher Models, Repositories, and Services
- Updated Address model to use defaultAddress instead of isDefault.
- Refactored CustomerVoucher model to remove isDeleted and isUsed fields.
- Modified Feedback model to link to Food instead of OrderItem.
- Changed Voucher model to use String code and updated related fields.
- Created FavoriteFoodRepository and FoodRepository for menu management.
- Updated MenuRepository to use Integer as ID type.
- Removed RecipeDetailRepository and RecipeRepository.
- Adjusted FoodTableRepository to use Integer as ID type.
- Simplified SalaryHistoryRepository by removing isDeleted checks.
- Cleaned up StaffRepository by removing isDeleted methods.
- Refined AddressRepository to focus on userId without isDeleted.
- Enhanced CustomerVoucherRepository with pagination support.
- Created AuditorAwareService for auditing purposes.
- Removed MenuItemService as it was not implemented.
- Expanded MenuService with methods for managing menus and foods.
- Updated FoodTableService to manage coffee table status and deletion.
- Refined OrderService to streamline order creation and management.
- Cleaned up StaffService to remove deleted logic and improve salary calculation.
- Removed AddressService and integrated its functionality into CustomerService.
- Created VoucherService for managing vouchers and customer interactions.
- Updated application.yml to comment out ddl-auto update for database management.

+++ server/src/main/java/io/foodapp/server/repositories/Menu/MenuRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import io.foodapp.server.models.MenuModel.Menu;
 
 @Repository
-public interface MenuRepository extends JpaRepository<Menu, Long> {
+public interface MenuRepository extends JpaRepository<Menu, Integer> {
     
 }
